GraphLIME
Describes a method for local interpretation of graph neural networks using a non-linear feature selection method. Specifically, create an N-hop local neighborhood for each node, and uses it to explain prediction for that specific node. Seems to work really well, but only focuses on prediction for a single target node - it may become much more relevant if the method can be extended to explaining a set of nodes (as in the case of a predictions based off chemical graphs)