Update README.md
First new thing for me in this paper was Extended Connectivity Finger Prints (ECFR), which seem to a provide a good basis for interpretation, in this case with integrated gradients, but I'm guessing one could use alternatives such as SHAP as well.

For graph convolutional interpretation, predictions are made directly from the node vectors (prior to pooling) to predict property, claiming that e.g. substructures (reflected in top vectors) predicting close to 1 are associated to positive labels, while those close to zero to negative labels. Interesting how this would generalize to regression tasks. As well as whether it generalizes at all.