Update README.md
Re-reading BioBERT with a better understanding of the NLP field. Not too much to say, as it's pretty much just BERT trained on PubMed/PMC, which naturally works significantly better than BERT on various tasks related to biomedical texts. Primary use cases for our current work are likely relation extraction and perhaps named entity recognition. I'd be very interested to know how performance relates to the size of the fine-tuning datasets, and thereby if more data is likely to help these tasks.