Update README.md
Perspectives on the future of AI in terms of creating hybrid model where neural network based learning is no longer the only first class citizen. While I couldn't agree more that we are not even close to where we need to be, and that hybrid/new methods and technologies are needed, I did feel like the paper is a bit.. unsympathetic.. to things such as e.g. model based RL, bayesian networks, and general work on causality.  Anyways, it was a really good read, and I indeed do look forward to watching this path of progress during the next decade.