Update README.md
Very important paper for the work I'm involved with right now, showing how a common attribution model (integrated gradients) used to interpret predictions made on molecular structures can reveal that these models capture spurious effects and biases in the data rather than the actual binding mechanism. They suggest an attribution AUC, which could be evaluated on a set of hard-coded binding rules, to see if such biases exist in a given dataset. Evaluating like this on synthetic targets is of course not ideal, but it's probably true that if the attribution AUC for such synthetic labels is low, then something is amiss with the dataset. We should definitely work further with the in our work.