Update README.md
Yet another approach for explaining predictions from graph neural networks. This one comes out of Stanford, and works bvy learning a mask of the adjacency matrix of the computation graph, and a feature mask for node features, while minimizing the uncertainty of the prediction made by the model (by minimizing conditional entropy conditioned on subgraphs and feature dimensions). Luckily the paper comes with the torch code, which is nice in terms of testing things out. Main focus of paper is node classification, but it also extends to edge or graph classification. Could well be worth including in current work to test out against alternative approaches.