Update README.md
Extremely interesting paper where they use causal dilated CNNs with attention to find temporal causalities in a set of timeseries. It does this by by using multiple CNNs, where each receives all time series and input and tries to predict one of the inputs. The internal parameters of the networks can subsequently be used for causal dicovery, specifically through an attention mechanism that allows us to identify which timeseries are being used. Subsequently they employ a causal validation step, where they intervene with the timeseries to test if it is actually causally related to the predicted timeseries.