Update README.md
Did not have time to thoroughly get through the math of this one, and I think it'd be beneficial to also go through some of the prior papers to fully grasp it. The overall idea of integrating steerable filters, group convolutions, and a densely connected framework to create a superior form of convolution which is more rotation equivariant seems brilliant, and is definitely something work looking into in the future when faced with a relevant dataset. Implementation by authors is in TF1, but hopefully someone will re-implement in TF2 or PyTorch given the encouraging results presented in this paper.