Update README.md
Paper is not out yet, but I've scavenged all the material put out so far, and I'm looking very much forward to diving into the paper when it comes out. It can be a little hard to really digest how the model works, but it's interesting that they seem to have ditched the pairwise distance plots as a driver of training, and instead output directly after the attention layers to the "structure module". It is mentioned that there is some iterative process going on, which I'm also not sure I completely understand; i.e. is there a feedback from the structure module back through the network. Very cool results for sure though.