Update README.md
Old but good for the basics - introduces reinforcement learning for trading, also explaining the differential sharpe ratio which can be used for online learning. I did not pick up the referenced papers for recurrent reinforcement learning; I want to read a few newer papers before going that far back in the literature to look for RL algorithms.