Update README.md
An interesting paper from deepmind about using probability trees as an alternative to bayesian networks for creating causal models. The advantage is that such models can represent context-specific causal dependencies that cannot be expressed within a bayesian network. They provide an open source colab, which makes it easy to grasp how to setup these models. Automatically learning the specific structure of the probabilty tree isn't addressed, and maybe it can't, so it does take some effort to use irl.