Update README.md
A practical approach to applying modern RL agents to stock trading, with pretty good backtesting performance. Introduces several RL agents, and use an ensemble of them in a live trading environment (i.e. regularly picking the best of them based on Sharpe). Code is given, as well as a nice medium article which also introduces the ideas and concepts nicely.