Update README.md
It wasn't quite clear whether this model should be put under traditional ML, interpretability, or new "tabular ML", so I placed it in interpretability, since it looks to be a super efficient transformer architecture for tabulat data, which also enables interpretability + the ability for self-supervised & online learning. The fact that this is on par or better than ensemble tree based methods is also super cool - maybe the days for me using lgbm, xgboost etc. are soon over.