Update README.md
This paper shows that we can improve image classification by increasing width, depth or resolution separately, but that the balancing of tuning them at the same time can dramatically improve scaling performance, which likely happens because the compound scaling enables the network to better focus on relevant features in the image. This drive towards more efficient architectures is extremely nice to see, as a counterpart to models growing more and more massive.