Update README.md
Rainbow is an ablation study of various improvements to traditional deep Q network RL agent, where it's found that combining 6 different individual improvements to the algorithm results in state of the art performance both in terms of data efficiency and final performance. Very nice paper, and I foresee that rainbow is definitely an algorithm I'll end up using in some context sooner or later.