Update README.md
I've been using the fast.ai library regularly and heard plenty of talks from Jeremy, but it was nevertheless nice to read this paper, to reiterate the design decisions behind the library, as well as fully get the type dispatch system and semantic tensors implemented in fastai. And also to re-iterate that fasiai is very hackable, whereas I've usually just been using it as a premade tool, and whenever I had to make custom stuff, I've done it from scratch.