Update README.md
A BERT model trained on SMILES strings, but interestingly using physchem properties extracted from Rdkit to learn more valuable representations, which work better for downstream tasks; this is a great idea, that seems to work quite well. I'm wondering, since attention is basically a graph operation, whether it's really necessary to go through the SMILES strings, but nevertheless this is nice & easy to read paper.