Update README.md
An interesting idea of using laplacian pooling for graph neural networks, which seem to improve interpretability, predictive and generative power of GNNs. Code is made available, which is very nice. Could be useful to include in some GNN work. Laplacian pooling, in contrast to alternative pooling methods, is suggested to be better at capturing relative importance of interactions between molecular substructures.