Update README.md
Another paper looking at some details related to GNN, specifically the used aggregation function, which they show is extremely important in terms of distinguising different messages passed in especially more complex graphs. This could be super interesting for a lot of GNN work, and they also do some interesting architectures where the number of layers depends on the number of nodes in the input graph. Code is made available and uses pytorch geometric, which is wonderful 👍