Update README.md
Deep kernel learning is a new concept for me, but definitely something I wish to read up on more. This paper is a pretty good introduction into the concept, and they introduce an task-adaptive kernel family for few-shot regression setting. It uses GP as the final regressor, and they also show very good performance when it comes to active learning, and so this could be extremely interesting to apply in industrial setting I think.