Update README.md
A generative approach for small molecule design, where the generative process is a Markov decision process of chemical reactions, which is pretty clever in terms of the suggested compounds being synthetically accessible (with the routes available in the model as well). Seems to perform really well, and in general be a solid approach towards generative model. Sadly I've not been able to find any released code for their REACTOR RL agent.